Python WEB/HW14-WEB/pyproject.toml
[tool.poetry]
name = "hw14-web"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "hw14_web"}]

[tool.poetry.dependencies]
python = "^3.10"
redis = "^4.5.5"
fastapi = "^0.97.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
libgravatar = "^1.0.4"
fastapi-mail = "^1.2.8"
redis-lru = "^0.1.2"
fastapi-limiter = "^0.1.5"
cloudinary = "^1.33.0"
sqlalchemy = "^1.3"
psycopg2 = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
